refactor(app): split AppDelegate into runtime, brain composition, and session artifacts
Closes #209. Completes Phase 5 of wiki/lean-coaching-core.md.

The 1,296-line delegate becomes three focused owners.

`AppDelegate` is the session runtime: Start, Stop, teardown, readiness
observation and its effects, capture-heartbeat handling, transcription
connection state, error reporting, hotkeys, the menu bar, and Settings
composition. It is also the host the other two report through.

`SessionArtifacts` owns everything a session leaves on disk: the
owner-only directory and its id, the log-directory resolution, the
evidence handle installed in it, the Activity projection's session,
jlog's attachment, retention pruning, the close bookkeeping that keeps a
still-sealing session from being pruned or evaluated, and the
source-checkout lookup evaluation needs.

`BrainComposition` owns provider preflight, brain-client construction,
route construction with its route-health callbacks, the live reapply of
brain preferences and credentials, and the route identity a Settings
edit is announced against.

`BrainCompositionHost` is the whole interface between composition and
the runtime: four read-only accessors for the live session and two
presentation forwards. Composition never starts, stops, or tears
anything down, and the runtime never builds a brain client.

Structure changed; observable lifecycle behavior did not — teardown
ordering, the Quit abandon path, and the closing-directory protection
are preserved exactly. check-ghost-mode.sh passes with no new
ghost-mode-allowed exceptions; both retained presentation calls keep
their inline reasons in the runtime, and neither extracted owner
presents anything.

Gate: swift build && ./scripts/run-tests.sh — 814 tests in 96 suites passed.
